refactor: Rename component using in WorkoutStudy
