+++ src/components/main/WorkoutStudyListItem.tsx
